+++ src/data/earthengine/s3.py
@@ -32,4 +32,5 @@ def get_single_s3_image(region: ee.Geometry, start_date: date, end_date: date) -
         return create_placeholder(region, S3_BANDS).toDouble()
 
     # all imagery has to have the same data type to be compatible
-    return image.toDouble()
+    #return image.toDouble()
+    return image
